pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "loomos"
version = "1.0.0"
description = "The Iron Suit for AI Models - Enterprise Distributed AI Runtime Platform"
authors = [
    {name = "Loom Labs", email = "hello@loomlabs.ai"},
    {name = "LoomOS Team", email = "team@loomlabs.ai"}
]
license = {text = "MIT"}
readme = "README.md"
homepage = "https://loomlabs.ai"
repository = "https://github.com/loomlabs/loomos"
documentation = "https://docs.loomlabs.ai"
keywords = [
    "ai", "machine-learning", "distributed-computing", "reinforcement-learning",
    "nexus", "training", "orchestration", "enterprise", "kubernetes"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Distributed Computing",
]
dependencies = [
    # Core framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    
    # Infrastructure & Orchestration
    "docker>=6.1.0",
    "kubernetes>=28.1.0",
    "boto3>=1.34.0",
    "nats-py>=2.6.0",
    "redis[hiredis]>=5.0.0",
    
    # Monitoring & Observability
    "prometheus-client>=0.19.0",
    "structlog>=23.2.0",
    
    # AI & Machine Learning
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "peft>=0.7.0",
    "accelerate>=0.25.0",
    "datasets>=2.16.0",
    
    # Reinforcement Learning
    "gymnasium>=0.29.0",
    "stable-baselines3>=2.2.0",
    "wandb>=0.16.0",
    
    # Distributed Computing
    "ray[default]>=2.8.0",
    "mpi4py>=3.1.0",
    
    # Security & Cryptography
    "cryptography>=41.0.0",
    "pyjwt>=2.8.0",
    
    # CLI & Utilities
    "typer[all]>=0.9.0",
    "rich>=13.7.0",
    "click>=8.1.0",
    
    # HTTP & Networking
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    
    # Data Processing
    "numpy>=1.24.0",
    "pandas>=2.1.0",
    "msgpack>=1.0.0",
    
    # Development & Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
]

enterprise = [
    "stripe>=7.8.0",
    "celery[redis]>=5.3.0",
    "flower>=2.0.0",
    "sentry-sdk[fastapi]>=1.39.0",
]

[project.urls]
Homepage = "https://loomlabs.ai"
Documentation = "https://docs.loomlabs.ai"
Repository = "https://github.com/loomlabs/loomos"
"Bug Reports" = "https://github.com/loomlabs/loomos/issues"
Changelog = "https://github.com/loomlabs/loomos/blob/main/CHANGELOG.md"

[project.scripts]
loomos = "cli.main:app"
nexus-cluster = "bin.nexus_cluster:main"

[tool.setuptools]
packages = ["loomos"]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"